# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Hello-books API",
    "description": "### Introduction\n API Documentation for the Hello-books API v1.0\n### Overview\nHello-Books is a simple application that helps manage a library and its processes like stocking, tracking and renting books. It is built with NodeJS/ExpressJS for the back-end and React, Redux and MaterializeCSS for the front-end.\n### Status Codes\n200 - A resource was retrieved successfully (e.g. Getting all books in the application)\n\n201 - A resource was created successfully (e.g. Adding a new book, Signing up a user)\n\n202 - Request has been sent and has a status of pending\n\n204 - Resource has been successfully deleted\n\n400 - An invalid request (e.g Submitting an invalid data type in the request)\n\n401 - Authentication failed, could arise from expired token or invalid credentials\n\n403 - Authorization failed. It could occur when a user tries to access privileged resources without admin token.\n\n404 - A resource was not found. It could occur when a user tries to get a book that does not exist.\n\n### Authorization\nPrefix assignrd token with `Bearer` when making a request to an authenticated endpoint."
  },
  "basePath": "/api/v1",
  "securityDefinitions": {
    "adminToken": {
      "description": "JWT token assigned to an admin",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    },
    "userToken": {
      "description": "JWT token assigned to a user",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/books": {
      "get": {
        "description": "",
        "summary": "Get books",
        "operationId": "BooksGet2",
        "tags": [
          "Books"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "sort",
            "in": "query",
            "type": "string",
            "description": "\"upvotes\" or \"downvotes\""
          },
          {
            "name": "order",
            "in": "query",
            "type": "string",
            "description": "\"desc\" or \"asc\""
          },
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "description": "The page to be retrieved"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "description": "The limit of books to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/GetBooksresponse"
            },
            "examples": {
              "application/json": {
                "message": "Successful",
                "books": [
                  {
                    "id": 96,
                    "title": "Love and Ruin",
                    "author": "Paul McClain",
                    "description": "Magnolia Table is infused with Joanna Gaines' warmth and passion for all things family, prepared and served straight from the heart of her home, with recipes inspired by dozens of Gaines family favorites and classic comfort selections from the couple's new Waco restaurant, Magnolia Table",
                    "subject": "Romance",
                    "imageURL": "https://images-na.ssl-images-amazon.com/images/I/51keKmj6MSL._SX327_BO1,204,203,200_.jpg",
                    "quantity": 20,
                    "borrowCount": 1,
                    "favCount": 0,
                    "upvotes": 1,
                    "downvotes": 0,
                    "createdAt": "2018-05-10T03:11:52.181Z",
                    "updatedAt": "2018-06-22T12:22:48.895Z",
                    "bookReviews": []
                  },
                  {
                    "id": 1,
                    "title": "Lean In",
                    "author": "Sheryl Sandberg",
                    "description": "A guide on how women can balance career and family",
                    "subject": "Self-help",
                    "imageURL": "https://res.cloudinary.com/ama-hello-books-v2/image/upload/v1529657777/Lean_In__book_cr3jst.jpg",
                    "quantity": 20,
                    "borrowCount": 0,
                    "favCount": 0,
                    "upvotes": 1,
                    "downvotes": 0,
                    "createdAt": "2018-06-22T08:56:20.614Z",
                    "updatedAt": "2018-06-22T12:22:28.308Z",
                    "bookReviews": []
                  }
                ],
                "pagination": {
                  "pageCount": 1,
                  "pageSize": 2,
                  "currentPage": 1,
                  "bookCount": 2
                }
              }
            }
          }
        },
        "x-unitTests": [
          {
            "request": {
              "method": "GET",
              "uri": "/books?sort=upvotes&order=desc"
            },
            "expectedResponse": {
              "x-allowExtraHeaders": true,
              "x-bodyMatchMode": "RAW",
              "x-arrayOrderedMatching": false,
              "x-arrayCheckCount": false,
              "x-matchResponseSchema": true,
              "headers": {
                "Connection": "keep-alive",
                "Content-Length": "1171",
                "Content-Type": "application/json; charset=utf-8",
                "Date": "Sat, 23 Jun 2018 19:37:55 GMT",
                "ETag": "W/\"493-bSaYQp4ftL9/RIG4twG/W2FvS4o\"",
                "X-Powered-By": "Express"
              },
              "body": "{\"message\":\"Successful\",\"books\":[{\"id\":96,\"title\":\"Love and Ruin\",\"author\":\"Paul McClain\",\"description\":\"Magnolia Table is infused with Joanna Gaines' warmth and passion for all things family, prepared and served straight from the heart of her home, with recipes inspired by dozens of Gaines family favorites and classic comfort selections from the couple's new Waco restaurant, Magnolia Table\",\"subject\":\"Romance\",\"imageURL\":\"https://images-na.ssl-images-amazon.com/images/I/51keKmj6MSL._SX327_BO1,204,203,200_.jpg\",\"quantity\":20,\"borrowCount\":1,\"favCount\":0,\"upvotes\":1,\"downvotes\":0,\"createdAt\":\"2018-05-10T03:11:52.181Z\",\"updatedAt\":\"2018-06-22T12:22:48.895Z\",\"bookReviews\":[]},{\"id\":1,\"title\":\"Lean In\",\"author\":\"Sheryl Sandberg\",\"description\":\"A guide on how women can balance career and family\",\"subject\":\"Self-help\",\"imageURL\":\"https://res.cloudinary.com/ama-hello-books-v2/image/upload/v1529657777/Lean_In__book_cr3jst.jpg\",\"quantity\":20,\"borrowCount\":0,\"favCount\":0,\"upvotes\":1,\"downvotes\":0,\"createdAt\":\"2018-06-22T08:56:20.614Z\",\"updatedAt\":\"2018-06-22T12:22:28.308Z\",\"bookReviews\":[]}],\"pagination\":{\"pageCount\":1,\"pageSize\":2,\"currentPage\":1,\"bookCount\":2}}"
            },
            "x-testShouldPass": true,
            "x-testEnabled": true,
            "x-testName": "Get books",
            "x-testDescription": ""
          }
        ],
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      },
      "post": {
        "description": "Endpoint for an authenticated admin to add a book",
        "summary": "Add a book",
        "operationId": "BooksPost",
        "tags": [
          "Books"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AddABookrequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/AddABookresponse"
            },
            "examples": {
              "application/json": {
                "message": "Successful",
                "bookEntry": {
                  "borrowCount": 0,
                  "favCount": 0,
                  "upvotes": 0,
                  "downvotes": 0,
                  "id": 5,
                  "title": "A new book",
                  "author": "Stephen Covey",
                  "description": "Dr. Covey's 7 Habits book is one of the most inspiring and impactful books ever written. Now you can enjoy and learn critical lessons about the habits of successful people that will enrich your life's experience",
                  "subject": "Inspiration",
                  "imageURL": "https://images-na.ssl-images-amazon.com/images/I/51fEYMhtHoL.jpg",
                  "quantity": 20,
                  "updatedAt": "2018-06-24T12:59:07.573Z",
                  "createdAt": "2018-06-24T12:59:07.573Z"
                }
              }
            }
          },
          "400": {
            "description": "The request cannot be fulfilled due to bad syntax.",
            "schema": {
              "$ref": "#/definitions/AddABookerrorresponse"
            },
            "examples": {
              "application/json": {
                "message": "Unsuccessful",
                "error": "ValidationError: child \"imageURL\" fails because [\"imageURL\" is required]"
              }
            }
          },
          "401": {
            "description": "Similar to 403 Forbidden, but specifically for use when authentication is possible but has failed or not yet been provided. The response must include a WWW-Authenticate header field containing a challenge applicable to the requested resource.",
            "schema": {
              "$ref": "#/definitions/AddABookerrorresponse"
            },
            "examples": {
              "application/json": {
                "message": "Unsuccessful",
                "error": "No token provided"
              }
            }
          },
          "403": {
            "description": "The request was a legal request, but the server is refusing to respond to it. Unlike a 401 Unauthorized response, authenticating will make no difference.",
            "schema": {
              "$ref": "#/definitions/AddABookerrorresponse"
            },
            "examples": {
              "application/json": {
                "message": "Unsuccessful",
                "error": "You must be an admin to access this feature"
              }
            }
          },
          "409": {
            "description": "Indicates that the request could not be processed because of conflict in the request, such as an edit conflict.",
            "schema": {
              "$ref": "#/definitions/AddABookerrorresponse"
            },
            "examples": {
              "application/json": {
                "message": "Unsuccessful",
                "error": "A book with this title already exists. Input a different title"
              }
            }
          }
        },
        "security": [
          {
            "adminToken": []
          }
        ],
        "x-unitTests": [
          {
            "request": {
              "method": "POST",
              "uri": "/books",
              "headers": {
                "Content-Type": "application/json"
              },
              "body": "{\"title\":\"A new book\",\"author\":\"Stephen Covey\",\"description\":\"Dr. Covey's 7 Habits book is one of the most inspiring and impactful books ever written. Now you can enjoy and learn critical lessons about the habits of successful people that will enrich your life's experience\",\"quantity\":20,\"subject\":\"Inspiration\"}"
            },
            "expectedResponse": {
              "x-allowExtraHeaders": true,
              "x-bodyMatchMode": "RAW",
              "x-arrayOrderedMatching": false,
              "x-arrayCheckCount": false,
              "x-matchResponseSchema": true,
              "headers": {
                "Connection": "keep-alive",
                "Content-Length": "568",
                "Content-Type": "application/json; charset=utf-8",
                "Date": "Sun, 24 Jun 2018 12:59:07 GMT",
                "ETag": "W/\"238-Fp98PesqZmrjpY79MNTJKw8/eK4\"",
                "X-Powered-By": "Express"
              },
              "body": "{\"message\":\"Successful\",\"bookEntry\":{\"borrowCount\":0,\"favCount\":0,\"upvotes\":0,\"downvotes\":0,\"id\":5,\"title\":\"A new book\",\"author\":\"Stephen Covey\",\"description\":\"Dr. Covey's 7 Habits book is one of the most inspiring and impactful books ever written. Now you can enjoy and learn critical lessons about the habits of successful people that will enrich your life's experience\",\"subject\":\"Inspiration\",\"imageURL\":\"https://images-na.ssl-images-amazon.com/images/I/51fEYMhtHoL.jpg\",\"quantity\":20,\"updatedAt\":\"2018-06-24T12:59:07.573Z\",\"createdAt\":\"2018-06-24T12:59:07.573Z\"}}"
            },
            "x-testShouldPass": true,
            "x-testEnabled": true,
            "x-testName": "Add a book",
            "x-testDescription": "Endpoint for an authenticated admin to add a book"
          }
        ],
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      }
    },
...
  },
  "definitions": {
...
    "Book15": {
      "title": "Book15",
      "example": {
        "id": 1,
        "title": "Lean In",
        "author": "Sheryl Sandberg",
        "description": "A guide on how women can balance career and family",
        "subject": "Self-help",
        "imageURL": "https://res.cloudinary.com/ama-hello-books-v2/image/upload/v1529657777/Lean_In__book_cr3jst.jpg",
        "quantity": 20,
        "borrowCount": 0,
        "favCount": 0,
        "upvotes": 1,
        "downvotes": 0,
        "createdAt": "2018-06-22T08:56:20.614Z",
        "updatedAt": "2018-06-22T12:22:28.308Z",
        "bookReviews": []
      },
      "type": "object",
      "properties": {
        "id": {
          "description": "",
          "example": 1,
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "description": "",
          "example": "Lean In",
          "type": "string"
        },
        "author": {
          "description": "",
          "example": "Sheryl Sandberg",
          "type": "string"
        },
        "description": {
          "description": "",
          "example": "A guide on how women can balance career and family",
          "type": "string"
        },
        "subject": {
          "description": "",
          "example": "Self-help",
          "type": "string"
        },
        "imageURL": {
          "description": "",
          "example": "https://res.cloudinary.com/ama-hello-books-v2/image/upload/v1529657777/Lean_In__book_cr3jst.jpg",
          "type": "string"
        },
        "quantity": {
          "description": "",
          "example": 20,
          "type": "integer",
          "format": "int32"
        },
        "borrowCount": {
          "description": "",
          "example": 0,
          "type": "integer",
          "format": "int32"
        },
        "favCount": {
          "description": "",
          "example": 0,
          "type": "integer",
          "format": "int32"
        },
        "upvotes": {
          "description": "",
          "example": 1,
          "type": "integer",
          "format": "int32"
        },
        "downvotes": {
          "description": "",
          "example": 0,
          "type": "integer",
          "format": "int32"
        },
        "createdAt": {
          "description": "",
          "example": "6/22/2018 8:56:20 AM",
          "type": "string"
        },
        "updatedAt": {
          "description": "",
          "example": "6/22/2018 12:22:28 PM",
          "type": "string"
        },
        "bookReviews": {
          "description": "",
          "example": [],
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "required": [
        "id",
        "title",
        "author",
        "description",
        "subject",
        "imageURL",
        "quantity",
        "borrowCount",
        "favCount",
        "upvotes",
        "downvotes",
        "createdAt",
        "updatedAt",
        "bookReviews"
      ]
    },
...
    "GetBooksresponse": {
      "title": "Get books",
      "example": {
        "message": "Successful",
        "books": [
          {
            "id": 96,
            "title": "Love and Ruin",
            "author": "Paul McClain",
            "description": "Magnolia Table is infused with Joanna Gaines' warmth and passion for all things family, prepared and served straight from the heart of her home, with recipes inspired by dozens of Gaines family favorites and classic comfort selections from the couple's new Waco restaurant, Magnolia Table",
            "subject": "Romance",
            "imageURL": "https://images-na.ssl-images-amazon.com/images/I/51keKmj6MSL._SX327_BO1,204,203,200_.jpg",
            "quantity": 20,
            "borrowCount": 1,
            "favCount": 0,
            "upvotes": 1,
            "downvotes": 0,
            "createdAt": "2018-05-10T03:11:52.181Z",
            "updatedAt": "2018-06-22T12:22:48.895Z",
            "bookReviews": []
          },
          {
            "id": 1,
            "title": "Lean In",
            "author": "Sheryl Sandberg",
            "description": "A guide on how women can balance career and family",
            "subject": "Self-help",
            "imageURL": "https://res.cloudinary.com/ama-hello-books-v2/image/upload/v1529657777/Lean_In__book_cr3jst.jpg",
            "quantity": 20,
            "borrowCount": 0,
            "favCount": 0,
            "upvotes": 1,
            "downvotes": 0,
            "createdAt": "2018-06-22T08:56:20.614Z",
            "updatedAt": "2018-06-22T12:22:28.308Z",
            "bookReviews": []
          }
        ],
        "pagination": {
          "pageCount": 1,
          "pageSize": 2,
          "currentPage": 1,
          "bookCount": 2
        }
      },
      "type": "object",
      "properties": {
        "message": {
          "example": "Successful",
          "type": "string"
        },
        "books": {
          "example": [
            {
              "id": 96,
              "title": "Love and Ruin",
              "author": "Paul McClain",
              "description": "Magnolia Table is infused with Joanna Gaines' warmth and passion for all things family, prepared and served straight from the heart of her home, with recipes inspired by dozens of Gaines family favorites and classic comfort selections from the couple's new Waco restaurant, Magnolia Table",
              "subject": "Romance",
              "imageURL": "https://images-na.ssl-images-amazon.com/images/I/51keKmj6MSL._SX327_BO1,204,203,200_.jpg",
              "quantity": 20,
              "borrowCount": 1,
              "favCount": 0,
              "upvotes": 1,
              "downvotes": 0,
              "createdAt": "2018-05-10T03:11:52.181Z",
              "updatedAt": "2018-06-22T12:22:48.895Z",
              "bookReviews": []
            },
            {
              "id": 1,
              "title": "Lean In",
              "author": "Sheryl Sandberg",
              "description": "A guide on how women can balance career and family",
              "subject": "Self-help",
              "imageURL": "https://res.cloudinary.com/ama-hello-books-v2/image/upload/v1529657777/Lean_In__book_cr3jst.jpg",
              "quantity": 20,
              "borrowCount": 0,
              "favCount": 0,
              "upvotes": 1,
              "downvotes": 0,
              "createdAt": "2018-06-22T08:56:20.614Z",
              "updatedAt": "2018-06-22T12:22:28.308Z",
              "bookReviews": []
            }
          ],
          "type": "array",
          "items": {
            "$ref": "#/definitions/Book15"
          }
        },
        "pagination": {
          "$ref": "#/definitions/Pagination"
        }
      },
      "required": [
        "message",
        "books",
        "pagination"
      ]
    },
...
    "AddABookrequest": {
      "title": "Add a book Request",
      "example": {
        "title": "A new book",
        "author": "Stephen Covey",
        "description": "Dr. Covey's 7 Habits book is one of the most inspiring and impactful books ever written. Now you can enjoy and learn critical lessons about the habits of successful people that will enrich your life's experience",
        "quantity": 20,
        "subject": "Inspiration",
        "imageURL": "https://images-na.ssl-images-amazon.com/images/I/51ciUK5JaCL._SX376_BO1,204,203,200_.jpg"
      },
      "type": "object",
      "properties": {
        "title": {
          "example": "A new book",
          "type": "string"
        },
        "author": {
          "example": "Stephen Covey",
          "type": "string"
        },
        "description": {
          "example": "Dr. Covey's 7 Habits book is one of the most inspiring and impactful books ever written. Now you can enjoy and learn critical lessons about the habits of successful people that will enrich your life's experience",
          "type": "string"
        },
        "quantity": {
          "example": 20,
          "type": "integer",
          "format": "int32"
        },
        "subject": {
          "description": "",
          "example": "Inspiration",
          "type": "string"
        },
        "imageURL": {
          "example": "https://images-na.ssl-images-amazon.com/images/I/51fEYMhtHoL.jpg",
          "type": "string"
        }
      },
      "required": [
        "title",
        "author",
        "description",
        "quantity",
        "subject",
        "imageURL"
      ]
    },
    "AddABookerrorresponse": {
      "title": "Add a book ErrorResponse",
      "example": {
        "message": "Unsuccessful",
        "error": "ValidationError: child \"imageURL\" fails because [\"imageURL\" is required]"
      },
      "type": "object",
      "properties": {
        "message": {
          "description": "",
          "example": "Unsuccessful",
          "type": "string"
        },
        "error": {
          "description": "",
          "example": "A book with this title already exists. Input a different title",
          "type": "string"
        }
      },
      "required": [
        "message",
        "error"
      ]
    },
    "AddABookresponse": {
      "title": "Add a book Response",
      "example": {
        "message": "Successful",
        "bookEntry": {
          "borrowCount": 0,
          "favCount": 0,
          "upvotes": 0,
          "downvotes": 0,
          "id": 5,
          "title": "A new book",
          "author": "Stephen Covey",
          "description": "Dr. Covey's 7 Habits book is one of the most inspiring and impactful books ever written. Now you can enjoy and learn critical lessons about the habits of successful people that will enrich your life's experience",
          "subject": "Inspiration",
          "imageURL": "https://images-na.ssl-images-amazon.com/images/I/51fEYMhtHoL.jpg",
          "quantity": 20,
          "updatedAt": "2018-06-24T12:59:07.573Z",
          "createdAt": "2018-06-24T12:59:07.573Z"
        }
      },
      "type": "object",
      "properties": {
        "message": {
          "description": "",
          "example": "Successful",
          "type": "string"
        },
        "bookEntry": {
          "$ref": "#/definitions/BookEntry"
        }
      },
      "required": [
        "message",
        "bookEntry"
      ]
    },
...
  }
}